// src/app/models/BD/usuario.model.ts
export class UsuarioModel {

     id_usuario: number;
     nombre: string;
     apellido: string;
     cedula: string;
     rol: number;
     usuario: string;
     contrasenia: string;
     puestotrabajo: string;
     telefono: string;
     obsercaion: string;
     id_favorito: number;
     imagen_perfil: string;
nstructor(
    ){}

}